fix: menu transactions to paginated menu
